ESP32 Synth: Comprehensive DSP overhaul and stability improvements
- Fully implemented Voice::process() signal chain including dual-oscillator mixing and selectable SVF/Ladder filtering.
- Re-enabled LFOs, Modulation Matrix, and global effects (Saturation, Chorus, Delay, Reverb, Compressor).
- Consolidated FM, Sync, and Ring modulation synthesis modes into the Voice class.
- Optimized Granular module with Hann window lookup tables and fast PRNG to reduce CPU load.
- Improved system stability by increasing task stack sizes and doubling DMA buffer size.
- Added AudioProfiler and SynthesisTests for real-time performance monitoring and DSP verification.
- Expanded MIDI CC mappings and PresetData structure for complete parameter persistence.
- Updated DisplayManager to show real-time voice activity and system health.
- Added tanhf soft-clipping to master output for audio safety.

+++ Granular.h
@@ -37,8 +37,8 @@ enum class GrainWindow : uint8_t {
 // Single grain state
 struct Grain {
     bool active;
-    float phase;
-    float phaseIncrement;
+    uint32_t phase;
+    uint32_t phaseIncrement;
     float amplitude;
     float position;     // 0-1 within grain duration
     float duration;     // In samples
